Refactor some misc code into util file
